+++ CHANGELOG.md
@@ -10,6 +10,7 @@ anymore.
 - 'Stop all' command now includes no auto components.
 - After 'stop all' a check for all components is triggered on the master server with broadcast option to inform all 
 connected clients.
+- Introduced default log level in config (set to INFO). This value is only overridden by the '--verbose' start parameter.
     
 ### Fixed
 - Error on shutdown before config was finished where no master node for dependency was found now is handled.